Refactor ES settings into environment.py; Pass proper es env vars into backend and huey containers
