+++ docker-compose.yaml
@@ -29,6 +29,13 @@ services:
       - API_BASE_URL=${API_BASE_URL}
       - APP_BASE_URL=${APP_BASE_URL}
       - PATH_TO_GEOIP_DB=${PATH_TO_GEOIP_DB}
+      - ENV=${ENV}
+      - PINECONE_API_KEY=${PINECONE_API_KEY}
+      - ELASTICSEARCH_URL=${ELASTICSEARCH_URL}
+      - ELASTICSEARCH_USER=${ELASTICSEARCH_USER}
+      - ELASTICSEARCH_PASSWORD=${ELASTICSEARCH_PASSWORD}
+      - ELASTICSEARCH_VERIFY_CERTS=${ELASTICSEARCH_VERIFY_CERTS}
+      - ELASTICSEARCH_INDEX_PREFIX=${ELASTICSEARCH_INDEX_PREFIX}
     ports:
       - "8000:8000"
     volumes:
@@ -64,6 +71,13 @@ services:
       - PATH_TO_GEOIP_DB=${PATH_TO_GEOIP_DB}
       - POSTMARK_API_KEY=${POSTMARK_API_KEY}
       - DEFAULT_SENDER_EMAIL=${DEFAULT_SENDER_EMAIL}
+      - PINECONE_API_KEY=${PINECONE_API_KEY}
+      - ENV=${ENV}
+      - ELASTICSEARCH_URL=${ELASTICSEARCH_URL}
+      - ELASTICSEARCH_USER=${ELASTICSEARCH_USER}
+      - ELASTICSEARCH_PASSWORD=${ELASTICSEARCH_PASSWORD}
+      - ELASTICSEARCH_VERIFY_CERTS=${ELASTICSEARCH_VERIFY_CERTS}
+      - ELASTICSEARCH_INDEX_PREFIX=${ELASTICSEARCH_INDEX_PREFIX}
     volumes:
       - ./backend:/app
     command: python -m huey.bin.huey_consumer main.huey
